+++ tests/TelegramShareTest.php
@@ -17,6 +17,18 @@ class TelegramShareTest extends TestCase
         $this->assertEquals($expected, (string)$result);
     }
 
+
+    /**
+     * @test
+     */
+    public function it_can_generate_a_telegram_share_instant_view_link_with_default_share_text()
+    {
+        $result = ShareFacade::page('https://codeswitch.be')->telegram(true, "606060");
+        $expected = '<div id="social-links"><ul><li><a target="_blank" href="https://t.me/iv?url=https://codeswitch.be&text=Default+share+text&rhash=606060" class="social-button " id="" title=""><span class="fab fa-telegram"></span></a></li></ul></div>';
+
+        $this->assertEquals($expected, (string)$result);
+    }
+
     /**
      * @test
      */
@@ -32,6 +44,18 @@ class TelegramShareTest extends TestCase
     /**
      * @test
      */
+    public function it_can_generate_a_telegram_share_instant_view_link_with_custom_share_text()
+    {
+        $result = ShareFacade::page('https://codeswitch.be', 'Meet Joren Van Hocht a php developer with a passion for laravel')
+            ->telegram(true, "606060");
+        $expected = '<div id="social-links"><ul><li><a target="_blank" href="https://t.me/iv?url=https://codeswitch.be&text=Meet+Joren+Van+Hocht+a+php+developer+with+a+passion+for+laravel&rhash=606060" class="social-button " id="" title=""><span class="fab fa-telegram"></span></a></li></ul></div>';
+
+        $this->assertEquals($expected, (string)$result);
+    }
+
+    /**
+     * @test
+     */
     public function it_can_generate_a_telegram_share_link_with_a_custom_class()
     {
         $result = ShareFacade::page('https://codeswitch.be', 'Meet Joren Van Hocht a php developer with a passion for laravel', ['class' => 'my-class'])
@@ -40,6 +64,7 @@ class TelegramShareTest extends TestCase
 
         $this->assertEquals($expected, (string)$result);
     }
+
 
     /**
      * @test